+++ src/yafowil/tests/test_common.py
@@ -3471,47 +3471,6 @@ class TestCommon(YafowilTestCase):
         data.write(model)
         self.assertEqual(model, {'URL': 'http://www.example.org'})
 
-    def test_search_blueprint(self):
-        # Render search input field
-        widget = factory(
-            'search',
-            name='SEARCH')
-        self.assertEqual(widget(), (
-            '<input class="search" id="input-SEARCH" name="SEARCH" '
-            'type="search" value="" />'
-        ))
-
-        # Extract not required and empty
-        data = widget.extract({'SEARCH': ''})
-        self.assertEqual(data.errors, [])
-
-        # Extract required empty
-        widget.attrs['required'] = True
-        data = widget.extract({'SEARCH': ''})
-        self.assertEqual(data.name, 'SEARCH')
-        self.assertEqual(data.value, UNSET)
-        self.assertEqual(data.extracted, '')
-        self.assertEqual(
-            data.errors,
-            [ExtractionError('Mandatory field was empty')]
-        )
-
-        del widget.attrs['required']
-
-        # Emptyvalue
-        widget.attrs['emptyvalue'] = 'defaultsearch'
-        data = widget.extract(request={'SEARCH': ''})
-        self.assertEqual(data.name, 'SEARCH')
-        self.assertEqual(data.value, UNSET)
-        self.assertEqual(data.extracted, 'defaultsearch')
-        self.assertEqual(data.errors, [])
-
-        data = widget.extract(request={'SEARCH': 'searchstr'})
-        self.assertEqual(data.name, 'SEARCH')
-        self.assertEqual(data.value, UNSET)
-        self.assertEqual(data.extracted, 'searchstr')
-        self.assertEqual(data.errors, [])
-
     def test_bytes_datatype_rendering_and_extraction(self):
         widget = factory(
             'text',
